# painter-2026-unuseable-ai/packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Painter 2026 API
  version: 1.0.0
  description: Contract-first APIs for painter-2026 services.
servers:
  - url: /api/v1
paths:
...
  /system/configs:
    get:
      summary: Fetch config center values by namespace
      operationId: listConfigs
      tags: [system]
      parameters:
        - in: query
          name: namespace
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Config list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseConfigListData'
...
  /system/admin/settings:
    get:
      summary: Admin get site settings
      operationId: getAdminSettings
      tags: [system]
      responses:
        '200':
          description: Site settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSettingsData'
    patch:
      summary: Admin update site settings
      operationId: updateAdminSettings
      tags: [system]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSettingsData'
components:
  schemas:
    ApiMeta:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        traceId: { type: string }
      required: [code, message, traceId]
...
    ApiResponseConfigListData:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              type: object
              properties:
                namespace: { type: string }
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ConfigItem'
              required: [namespace, items]
...
    ApiResponseSettingsData:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SiteSettings'
...
    ConfigItem:
      type: object
      properties:
        key: { type: string }
        value: { type: string }
        valueType: { type: string, enum: [string, number, boolean, json] }
        updatedAt: { type: string, format: date-time }
      required: [key, value, valueType, updatedAt]
...
    SiteSettings:
      type: object
      properties:
        siteName: { type: string }
        icpCode: { type: string }
        github: { type: string }
        canRegister: { type: boolean }
      required: [siteName, canRegister]
    UpdateSettingsRequest:
      type: object
      properties:
        siteName: { type: string }
        icpCode: { type: string }
        github: { type: string }
        canRegister: { type: boolean }